app/requirements.txt
PyQt6>=6.6.0
pyaudio>=0.2.14
google-genai>=1.0.0
openai>=1.40.0
pydub>=0.25.1
audioop-lts>=0.2.0; python_version>="3.13"
markdown>=3.5.0
pynput>=1.7.6
evdev>=1.6.0; sys_platform=="linux"
onnxruntime>=1.16.0
httpx>=0.27.0
requests>=2.31.0
mongita>=1.2.0
